feat: Complete endpoints, controllers, and tested with Postman
- Created all necessary controllers for the application
- Developed and tested endpoints using Postman to ensure functionality
- Refactored services for proper endpoint operation
- Fixed issues and ensured all tests pass

+++ src/main/java/org/example/ecommerce_api/Domain/Product/DataRegisterProduct.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 public record DataRegisterProduct(
         @NotNull String name,
         @NotNull Long typeProductId,
-        @NotNull BigDecimal price,
+        @NotNull Double price,
         @NotNull Integer rating,
         @NotNull String image,
         @NotNull String description,
